requirements.txt
streamlit~=1.37.0
streamlit-ldap-authenticator
python-dotenv~=1.0.1
langchain~=0.2.12
langchain-community
langchain-chroma
glpi-api
langchain-core
PyMuPDF
pydantic~=2.8.2
requests~=2.32.3
PyYAML~=6.0.1
ldap3~=2.9.1
beautifulsoup4~=4.12.3
tiktoken
langchain-openai
langchain-mistralai
mistralai~=1.0.1
azure-identity
langchain-community
